refactor: Implement layered architecture and lazy loading
