+++ plugin/uploadArticle/plugins/upload_to_all_platform.js
@@ -1,13 +1,14 @@
-const upload = require("./uploadUtil");
+const UploadUtil = require("../Plugin2UploadBridge");
 
-class uploadToAllPlatform extends upload {
+class uploadToAllPlatform extends BasePlugin {
 
     hotkey = () => [{hotkey: this.config.HOTKEY, callback: this.call}]
 
 
     call = () => {
+        this.baseUploader = new UploadUtil(this);
         let filePath = this.utils.getFilePath();
-        this.uploadProxy(filePath);
+        this.baseUploader.uploadProxy(filePath);
     }
 
 }
