+++ plugin/uploadArticle/请读我.md
@@ -1,6 +1,6 @@
 # 自动上传文章插件
 
-由于CSDN是直接逆向的接口，上传文章大概是2-3s，两个selenium自动化的平台上传时间长一点，20s左右
+由于CSDN是直接逆向的接口，上传文章大概是2s左右，两个selenium自动化登录的平台上传时间长一点，20s左右
 
 ### 原因
 
@@ -34,14 +34,6 @@
 
 ### 配置
 
-❗**版本提示**：插件使用的selenium谷歌驱动是125.0.2的，需要保证本机有Google Chrome。如果发布失败，需要检查版本号。方法：
-
-1. 你可以访问这个链接获取脚本匹配的谷歌浏览器版本[下载](https://dl.google.com/release2/chrome/acgeoucdi4yjn3qutvqsm727ugta_125.0.6422.61/125.0.6422.61_chrome_installer.exe) ，或者你可以访问[Release 自动发布博客依赖 ](https://github.com/AZCodingAccount/typora_plugin/releases/tag/1.8.46)进行下载
-
-2. 使用npm安装的同学你也可以修改package.json下方`chromedriver`版本号以匹配你的版本
-
->建议关闭自动更新，可能会有版本问题，但是我一直都是在更新半年了好像也没出现问题
-
 #### 引入依赖
 
 默认两个插件是启用的，但需要**引入一些依赖**：
@@ -62,7 +54,7 @@
 
 1. 如果你是程序开发人员，电脑有node环境，在当前路径下，运行`npm i`即可
 
-   ⚠️ 运行之前先把之前的`node_modules`包的几个依赖给备份一下（如下），下载好再粘回来。使用npm下载依赖会删除原有包
+   ⚠️ 运行之前先把之前的`node_modules`包的几个依赖给备份一下（如下），下载好再粘回来。使用node下载依赖会删除原有的`fswin`包
 
    <img src="https://my-picture-bed1-1321100201.cos.ap-beijing.myqcloud.com/mypictures/image-20240527175003616.png" alt="image-20240527175003616" style="zoom:50%;" />
 
@@ -90,13 +82,13 @@
 
 <img src="https://my-picture-bed1-1321100201.cos.ap-beijing.myqcloud.com/mypictures/image-20240527214113433.png" alt="image-20240527214113433" width=50% />
 
-#### 启用
+#### 禁用
 
-如果您想要**启用**，首先进入`.\plugin\global\settings`，找到`settings.user.toml`文件（初次使用需要先拷贝`settings.default.toml`一份，具体参考作者文档[typora_plugin/plugin/global/settings/请读我.md](https://github.com/obgnail/typora_plugin/blob/master/plugin/global/settings/请读我.md)。并看下方图片将注释放开即可
+如果您想要**禁用**，首先进入`.\plugin\global\settings`，找到`settings.user.toml`文件（初次使用需要先拷贝`settings.default.toml`一份，具体参考作者文档[typora_plugin/plugin/global/settings/请读我.md](https://github.com/obgnail/typora_plugin/blob/master/plugin/global/settings/请读我.md)。并看下方图片
 
 <img src="https://my-picture-bed1-1321100201.cos.ap-beijing.myqcloud.com/mypictures/%E7%A6%81%E7%94%A8.gif" alt="禁用" width=50% />
 
-
+<img src="https://my-picture-bed1-1321100201.cos.ap-beijing.myqcloud.com/mypictures/image-20240527200020108.png" alt="image-20240527200020108" width=50%/>
 
 
 
@@ -106,11 +98,21 @@
 
 ```mermaid
 graph TD
-    A[BasePlugin] --> B[uploadUtil]
+    B[BasePlugin]
     B --> C[uploadToCNBlogPlugin]
-    B --> D[uploadToCSDN]
-    B --> E[uploadToWordPress]
-    B --> F[uploadToAllPlatform]
+    B --> D[uploadToCSDNPlugin]
+    B --> E[uploadToWordPressPlugin]
+    B --> F[uploadToAllPlatformPlugin]
+    C --> G[Plugin2UploadBridge]
+    D --> G[Plugin2UploadBridge]
+    E --> G[Plugin2UploadBridge]
+    F --> G[Plugin2UploadBridge]
+    G --> H[UploadController]
+    H --> I[BaseUploaderInterface]
+    I --> J[CnBlogUploader]
+    I --> K[CsdnUploader]
+    I --> L[WordpressUploader]
+    
 ```
 
 #### CSDN
